+++ test/step_definitions/stepsCJS.js
@@ -1,0 +1,5 @@
+const { When } = require('@cucumber/cucumber');
+
+When('I am CJS', async function() {
+    console.log('CJS');
+});